check permissions of files that are created

+++ tests/end_to_end_tests/test_minishell_redirection.py
@@ -1,7 +1,30 @@
 import pytest
 import os
 import time
+import stat
 from conftest import start_process, get_prompt_minishell
+
+
+def check_permission(want_perm, file_path):
+    mode = os.stat(file_path).st_mode
+
+    user_perm = (
+        4 * bool(mode & stat.S_IRUSR)
+        + 2 * bool(mode & stat.S_IWUSR)
+        + bool(mode & stat.S_IXUSR)
+    )
+    group_perm = (
+        4 * bool(mode & stat.S_IRGRP)
+        + 2 * bool(mode & stat.S_IWGRP)
+        + bool(mode & stat.S_IXGRP)
+    )
+    other_perm = (
+        4 * bool(mode & stat.S_IROTH)
+        + 2 * bool(mode & stat.S_IWOTH)
+        + bool(mode & stat.S_IXOTH)
+    )
+    got_perm = 100 * user_perm + 10 * group_perm + other_perm
+    assert got_perm == want_perm
 
 
 @pytest.mark.parametrize(
@@ -25,14 +48,11 @@ def test_out_redirections(cmd):
 
     tmp_path = "tests/end_to_end_tests/test_files/non_existent.txt"
     assert os.path.isfile(tmp_path), f"file does not exist {tmp_path}"
-    while not os.access(tmp_path, os.R_OK):
-        time.sleep(0.1)
+    check_permission(644, tmp_path)
     with open(tmp_path, "r") as f:
         file_bash = f.readlines()
-
     os.remove(tmp_path)
-    while os.path.isfile(tmp_path):
-        time.sleep(0.1)
+    assert not os.path.isfile(tmp_path), f"file does still exist {tmp_path}"
 
     prompt = get_prompt_minishell()
     minishell = start_process("./minishell")
@@ -40,13 +60,11 @@ def test_out_redirections(cmd):
     stdout_minishell, stderr_minishell = minishell.communicate(cmds.encode())
 
     assert os.path.isfile(tmp_path), f"file does not exist {tmp_path}"
-    while not os.access(tmp_path, os.R_OK):
-        time.sleep(0.1)
+    check_permission(644, tmp_path)
     with open("tests/end_to_end_tests/test_files/non_existent.txt", "r") as f:
         file_minishell = f.readlines()
     os.remove("tests/end_to_end_tests/test_files/non_existent.txt")
-    while os.path.isfile(tmp_path):
-        time.sleep(0.1)
+    assert not os.path.isfile(tmp_path), f"file does still exist {tmp_path}"
 
     stdout_minishell = [
         line
